Register alembic_migrations at install time
This allows neutron-db-manage to find the alembic migrations
automatically if this project is installed.
The config file alembic.ini is not needed in this project since it's
tables are in neutron's DB and so it depends on neutron's config.
Remove the baremetal-networking-db-manage script since we will use
neutron-db-manage.
Update the alembic env to know about all neutron core tables when
generating diffs for --autogenerate.